LchemySQL now running with Postgresql (8.4), Mysql (5.1), sqlite. AlchemySqlModelTestCase subclassed to run Sqlite (as base class AlchemySqlModelTestCase), Posqresql (SqlaPostgresqlModelTestCase) and Mysql (SqlaMysqlModelTesCase). Explicit model.close() calls added to tests to preserve connection pool and mem resources - necessitated no-op close() method added to FileStore base class.

+++ src/vesper/data/store/alchemysql.py
@@ -4,12 +4,11 @@ __all__ = ['AlchemySQLStore']
 
 import os, os.path
 
-# MADNESS!
 import sqlalchemy
 from sqlalchemy import engine, sql, create_engine
-from sqlalchemy.sql import select
-from sqlalchemy.sql.expression import func, or_, and_, not_, exists, case
 from sqlalchemy.types import *
+from sqlalchemy.sql import select
+from sqlalchemy.sql.expression import func, or_, asc
 from sqlalchemy.schema import Table, Column, MetaData, UniqueConstraint, Index
 
 from vesper.backports import *
@@ -29,8 +28,7 @@ class AlchemySqlStore(Model):
       context UNIQUE
     )
     '''
-
-    def __init__(self, source = None, defaultStatements = None, autocommit=False, **kw):
+    def __init__(self, source=None, defaultStatements=None, pStore=False, autocommit=False, **kw):
         if source is None:
             # this seems like a reasonable default thing to do 
             source = 'sqlite://'
@@ -38,32 +36,29 @@ class AlchemySqlStore(Model):
 
         # We take source to be a SQLAlchemy-style dbapi spec: 
         # dialect+driver://username:password@host:port/database
-        # connection is made JIT on first connect()
         log.debug("sqla engine being created with:", source)
         self.engine = create_engine(source)
         self.md = sqlalchemy.schema.MetaData()
-        # utterly insufficient datatypes. just for first pass
-        # technically the keep_existing bool is redundant as create_all() default is "check first"
+        if not pStore:
+            self.engine.execute("drop table if exists vesper_stmts;")
+
+        # create our simple quin-tuple store
         self.vesper_stmts = Table('vesper_stmts', self.md, 
-                             Column('subject', String(255)),    # primary_key = True),
-                             Column('predicate', String(255)),  # primary_key = True),
-                             Column('object', String(255)),     # primary_key = True),
-                             Column('objecttype', String(8)),
-                             Column('context', String(8)),
-                             UniqueConstraint('subject', 'predicate', 'object', 'objecttype', 'context'),
-                             keep_existing = True)
+                                  Column('subject', String(255)),
+                                  Column('predicate', String(255)),
+                                  Column('object', String(255)),
+                                  Column('objecttype', String(8)),
+                                  Column('context', String(8)),
+                                  UniqueConstraint('subject', 'predicate', 'object', 'objecttype', 'context'),
+                                  mysql_engine='InnoDB', 
+                                  keep_existing = False)
         Index('idx_vs', self.vesper_stmts.c.subject, self.vesper_stmts.c.predicate, self.vesper_stmts.c.object) 
         self.md.create_all(self.engine)
-
-        # Set up our state machine and grab a connection from the sqlalchemy pool
-        self.conn = self.engine.connect()
+        
+        # attend to private matters of state
+        self.conn = None
         self.trans = None
         self.autocommit = autocommit
-
-        # XXX testing
-        result = self.conn.execute(select([func.count(self.vesper_stmts.c.subject)]))
-        for r in result: 
-            print "rows at init: ", r
 
     def _checkConnection(self):
         if self.conn is None:
@@ -100,13 +95,15 @@ class AlchemySqlStore(Model):
                 objecttype = OBJECT_TYPE_LITERAL
 
         if not asQuad and not fc:
-            query = select([self.vesper_stmts.c.subject, 
-                            self.vesper_stmts.c.predicate,
-                            self.vesper_stmts.c.object, 
-                            self.vesper_stmts.c.objecttype,
-                            func.min(self.vesper_stmts.c.context).label('context')])
-        else:        # asQuad is True
-            query = self.vesper_stmts.select()
+            query = select(['subject', 'predicate', 'object', 'objecttype',
+                            func.min(self.vesper_stmts.c.context).label('context')]).\
+                            group_by('subject', 'predicate', 'object', 'objecttype')
+        else:
+            if self.engine.name != 'postgresql':
+                query = self.vesper_stmts.select()
+            else:
+                query = self.vesper_stmts.select().\
+                    order_by('subject', 'predicate', 'object','objecttype', 'context')
         if fs:
             query = query.where(self.vesper_stmts.c.subject == subject)
         if fp:
@@ -117,14 +114,9 @@ class AlchemySqlStore(Model):
             query = query.where(self.vesper_stmts.c.objecttype == objecttype)
         if fc:
             query = query.where(self.vesper_stmts.c.context == context)
-        if not asQuad and not fc:
-            query = query.group_by(self.vesper_stmts.c.subject,
-                                   self.vesper_stmts.c.predicate,
-                                   self.vesper_stmts.c.object,
-                                   self.vesper_stmts.c.objecttype)
-        if limit is not None:
+        if limit:
             query = query.limit(limit)
-        if offset is not None:
+        if offset:
             query = query.offset(offset)
 
         stmts = []
@@ -132,7 +124,7 @@ class AlchemySqlStore(Model):
         result = self.conn.execute(query)
         for r in result:
             stmts.append( Statement(r['subject'], r['predicate'], r['object'], r['objecttype'], r['context']) )
-           
+            
         log.debug("stmts returned: ", len(stmts), stmts)
         return stmts
 
@@ -157,7 +149,7 @@ class AlchemySqlStore(Model):
         try:
             result = self.conn.execute(ins)
         except sqlalchemy.exc.IntegrityError, exc:
-            return 1    # or one?
+            return 0    # or one?
         return result.rowcount
 
     def addStatements(self, stmts):
@@ -173,17 +165,11 @@ class AlchemySqlStore(Model):
         ins = self.vesper_stmts.insert()
         self._checkConnection()
 
-        if self.engine.name == "postgresql":
-            result = self.conn.execute(
-"BEGIN LOOP BEGIN INSERT INTO vesper_stmts VALUES (%s, %s, %s, %s, %s); RETURN; \
- EXCEPTION WHEN unique_violation THEN END; END LOOP; END", 
-                argDictList)
-            return result.rowcount
-
-        elif self.engine.name == "sqlite":
+        if self.engine.name == "sqlite":
             ins = ins.prefix_with("OR IGNORE")
         elif self.engine.name == "mysql":
             ins = ins.prefix_with("IGNORE")
+
         result = self.conn.execute(ins, argDictList)
         return result.rowcount
 
@@ -201,7 +187,7 @@ class AlchemySqlStore(Model):
         result = self.conn.execute(rmv)
         return result.rowcount
 
-    def removeStatements(self, stmts):
+    def removeStatements(self, stmts=None):
         '''removes multiple statements from the model'''
         log.debug("removeStatements called with: ", stmts)
 
@@ -211,8 +197,11 @@ class AlchemySqlStore(Model):
                    (self.vesper_stmts.c.object == stmt[2]) & 
                    (self.vesper_stmts.c.objecttype == stmt[3]) & 
                    (self.vesper_stmts.c.context == stmt[4])) for stmt in stmts]
-        # no protection for singleton stmt here!
-        rmv = self.vesper_stmts.delete().where(or_(*wc))
+
+        rmv = self.vesper_stmts.delete()         
+        if stmts is not None:
+            rmv = self.vesper_stmts.delete().where(or_(*wc))
+            
         self._checkConnection()
         result = self.conn.execute(rmv)
         return result.rowcount
@@ -229,7 +218,6 @@ class AlchemySqlStore(Model):
 
     def close(self):
         log.debug("closing!")
-        print "closing!"
         if self.conn is not None:
             self.conn.close()
             self.conn = None
